fix: sar stake and sar positions duplication
