+++ src/state/pmulticall/hooks.ts
@@ -105,7 +105,7 @@ function useCallsData(calls: (Call | undefined)[], options?: ListenerOptions): C
   );
 }
 
-interface CallState {
+export interface CallState {
   readonly valid: boolean;
   // the result, or undefined if loading or errored/no data
   readonly result: Result | undefined;
