+++ src/state/psarstake/hooks.ts
@@ -4,21 +4,18 @@ import { TransactionResponse } from '@ethersproject/providers';
 import { JSBI, Token, TokenAmount } from '@pangolindex/sdk';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { BIGNUMBER_ZERO, ZERO_ADDRESS } from 'src/constants';
+import { BIGNUMBER_ZERO } from 'src/constants';
 import { PNG } from 'src/constants/tokens';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
-import { MixPanelEvents, useMixpanel } from 'src/hooks/mixpanel';
-import { useUSDCPriceHook } from 'src/hooks/multiChainsHooks';
-import { useApproveCallback } from 'src/hooks/useApproveCallback';
+import { MixPanelEvents } from 'src/hooks/mixpanel';
 import { useSarStakingContract } from 'src/hooks/useContract';
 import { calculateGasMargin, existSarContract, waitForTransaction } from 'src/utils';
 import { maxAmountSpend } from 'src/utils/maxAmountSpend';
 import { useSingleCallResult, useSingleContractMultipleData } from '../pmulticall/hooks';
-import { useDerivedStakeInfo } from '../pstake/hooks';
 import { tryParseAmount } from '../pswap/hooks';
 import { useTransactionAdder } from '../ptransactions/hooks';
-import { useTokenBalance } from '../pwallet/hooks';
 import { Position, URI } from './types';
+import { formatPosition, useDefaultSarStake } from './utils';
 
 // Return the info of the sar stake
 export function useSarStakeInfo() {
@@ -50,71 +47,33 @@ export function useSarStakeInfo() {
  * @returns Return some utils functions for stake more or create a new Position
  */
 export function useDerivativeSarStake(positionId?: BigNumber) {
-  const [attempting, setAttempting] = useState(false);
-  const [hash, setHash] = useState<string | null>(null);
-  const [typedValue, setTypedValue] = useState('');
-  const [stepIndex, setStepIndex] = useState(4);
-  const [stakeError, setStakeError] = useState<string | null>(null);
-
-  const { account } = usePangolinWeb3();
-  const chainId = useChainId();
-
-  const sarStakingContract = useSarStakingContract();
-
-  const addTransaction = useTransactionAdder();
-  const { t } = useTranslation();
-
-  const png = PNG[chainId];
-
-  const userPngBalance = useTokenBalance(account ?? ZERO_ADDRESS, png);
-
-  // used for max input button
-  const maxAmountInput = maxAmountSpend(chainId, userPngBalance);
-
-  const useUSDCPrice = useUSDCPriceHook[chainId];
-  const usdcPrice = useUSDCPrice(png);
-  const dollerWorth =
-    userPngBalance?.greaterThan('0') && usdcPrice ? Number(typedValue) * Number(usdcPrice.toFixed()) : undefined;
-
-  const wrappedOnDismiss = useCallback(() => {
-    setStakeError(null);
-    setTypedValue('');
-    setStepIndex(0);
-    setHash(null);
-    setAttempting(false);
-  }, []);
-
-  const { parsedAmount, error } = useDerivedStakeInfo(typedValue, png, userPngBalance);
-  const [approval, approveCallback] = useApproveCallback(chainId, parsedAmount, sarStakingContract?.address);
-
-  const mixpanel = useMixpanel();
-
-  const onUserInput = useCallback((_typedValue: string) => {
-    setTypedValue(_typedValue);
-  }, []);
-
-  const handleMax = useCallback(() => {
-    maxAmountInput && onUserInput(maxAmountInput.toExact());
-    setStepIndex(4);
-  }, [maxAmountInput, onUserInput]);
-
-  const onChangePercentage = (value: number) => {
-    if (!userPngBalance) {
-      setTypedValue('0');
-      return;
-    }
-    if (value === 100) {
-      setTypedValue((userPngBalance as TokenAmount).toExact());
-    } else if (value === 0) {
-      setTypedValue('0');
-    } else {
-      const newAmount = (userPngBalance as TokenAmount)
-        .multiply(JSBI.BigInt(value))
-        .divide(JSBI.BigInt(100)) as TokenAmount;
-
-      setTypedValue(newAmount.toSignificant(6));
-    }
-  };
+  const {
+    account,
+    addTransaction,
+    approval,
+    approveCallback,
+    attempting,
+    chainId,
+    dollerWorth,
+    error,
+    handleMax,
+    hash,
+    mixpanel,
+    onChangePercentage,
+    onUserInput,
+    parsedAmount,
+    png,
+    sarStakingContract,
+    setAttempting,
+    setHash,
+    setStakeError,
+    setStepIndex,
+    stakeError,
+    stepIndex,
+    t,
+    typedValue,
+    wrappedOnDismiss,
+  } = useDefaultSarStake();
 
   const onStake = async () => {
     if (!sarStakingContract || !parsedAmount) {
@@ -565,34 +524,12 @@ export function useSarPositions() {
       }
       return {} as URI;
     });
-    const positions: Position[] = nftsURIs.map((uri, index) => {
-      const valueVariables: { balance: BigNumber; sumOfEntryTimes: BigNumber } | undefined =
-        positionsAmountState[index].result?.valueVariables;
-      const rewardRate = positionsRewardRateState[index].result?.[0];
-      const pendingRewards = positionsPedingRewardsState[index].result?.[0];
-      const id = nftsIndexes[index][0];
-      const balance = valueVariables?.balance ?? BigNumber.from(0);
-      const apr = rewardRate
-        ?.mul(86400)
-        .mul(365)
-        .mul(100)
-        .div(balance.isZero() ? 1 : balance);
-
-      if (!valueVariables || !rewardRate || !pendingRewards || !uri) {
-        return {} as Position;
-      }
-
-      return {
-        id: BigNumber.from(id),
-        balance: valueVariables?.balance,
-        sumOfEntryTimes: valueVariables?.sumOfEntryTimes,
-        apr: apr,
-        rewardRate: rewardRate,
-        uri: uri,
-        pendingRewards: pendingRewards,
-      } as Position;
-    });
-    // remove the empty positions
-    return { positions: positions.filter((position) => !!position), isLoading: false };
+    return formatPosition(
+      nftsURIs,
+      nftsIndexes,
+      positionsAmountState,
+      positionsRewardRateState,
+      positionsPedingRewardsState,
+    );
   }, [account, positionsAmountState, positionsRewardRateState, nftsURIsState, nftsIndexes]);
 }
